Indexed db

+++ Frontend/src/views/GeneMapper/Home/index.jsx
@@ -73,18 +73,30 @@ function GeneMapperHome({ style, loggedIn }) {
   // function to handle deletion of project
   const handleDeleteProject = (id) => {
     DemoService.getDemos().then((demos) => {
-      ProjectService.deleteProject(id).then(() => {
-        ProjectService.getOwnProjects().then((data) => {
-          // search for demos and set the information stored about them
-          findDemos(data, demos);
-          setProjects(data);
+      // delete project from cache permanently
+      let cachedProjects = JSON.parse(localStorage.getItem("cached_projects"));
+      console.log(`Cached projects: ${JSON.stringify(cachedProjects)}`);
+      delete cachedProjects[id];
+      localStorage.setItem("cached_projects", JSON.stringify(cachedProjects));
+
+      // delete project stored on the remotely
+      if (loggedIn) {
+        // update the projects that are not deleted
+        ProjectService.deleteProject(id).then(() => {
+          ProjectService.getOwnProjects().then((data) => {
+            // search for demos and set the information stored about them
+            findDemos(data, demos);
+            setProjects(data);
+          });
+          // update the projects that are deleted
+          ProjectService.getDeletedProjects().then((data) => {
+            // search for demos and set the information stored about them
+            findDemos(data, demos);
+            console.log(data);
+            setDeletedProjects(data);
+          });
         });
-        ProjectService.getDeletedProjects().then((data) => {
-          // search for demos and set the information stored about them
-          findDemos(data, demos);
-          setDeletedProjects(data);
-        });
-      });
+      }
     });
   };
 
@@ -95,6 +107,7 @@ function GeneMapperHome({ style, loggedIn }) {
         ProjectService.getOwnProjects().then((data) => {
           // search for demos and set the information stored about them
           findDemos(data, demos);
+          // update the information about the demos
           setProjects(data);
         });
         ProjectService.getDeletedProjects().then((data) => {
@@ -106,10 +119,6 @@ function GeneMapperHome({ style, loggedIn }) {
     });
   };
 
-  // todo: store the projects in the cache as well
-  // in the cache, store an array of projects using the following object structure
-  // {_id, owner, name, modelId, atlasId, fileName, fileSize, status, resultSize, _v, uploadId, location}
-  // Function to get projects and update the necessary info about the demo datasets
   const getProjects = () => {
     // fetch demos
     DemoService.getDemos().then((demos) => {
@@ -117,7 +126,32 @@ function GeneMapperHome({ style, loggedIn }) {
         // search for demos and set the information stored about them
         findDemos(data, demos);
         setDemoDatasets(demos);
-        setProjects(data);
+
+        // check if indexedDB is available for caching the projects
+        if (!window.indexedDB) {
+          setProjects(data);
+          return;
+        }
+        // Save an object that consists of the projects in the local storage.
+        if (localStorage.getItem("cached_projects") === null) {
+          localStorage.setItem("cached_projects", JSON.stringify({}));
+        }
+        let cachedProjects = JSON.parse(localStorage.getItem("cached_projects"));
+
+        // filter out duplicate projects
+        data.forEach((obj) => {
+          let id = obj._id;
+          // If the cached projects object doesn't have a project with a matching id,
+          // add it to the cache. 
+          if (!cachedProjects.hasOwnProperty(id)) {
+            cachedProjects[id] = obj;
+          }
+        });
+        // turn cached project object to array and reverse the order
+        let projectArr = Object.keys(cachedProjects).map((key) => cachedProjects[key]).reverse();
+        setProjects(projectArr);
+        // update the local storage
+        localStorage.setItem("cached_projects", JSON.stringify(cachedProjects));
       });
     });
   };
@@ -136,7 +170,12 @@ function GeneMapperHome({ style, loggedIn }) {
     ModelService.getModels().then((data) => setModels(data));
     if (loggedIn) {
       TeamService.getMyTeams().then((teams) => setUserTeams(teams));
-      ProjectService.getDeletedProjects().then((data) => setDeletedProjects(data));
+      DemoService.getDemos().then((demos) => {
+        ProjectService.getDeletedProjects().then((data) => {
+          findDemos(data, demos);
+          setDeletedProjects(data);
+        });
+      });
     }
   }, []);
 
@@ -165,18 +204,18 @@ function GeneMapperHome({ style, loggedIn }) {
             />
           </Box>
         </Box>
-        {/* Check if jwt token exists to make sure that checking the projects is even necessary */}
-        {localStorage.getItem('jwt') && projects === null
+        {/* Check if project cache exists to make sure that checking the projects is even necessary */}
+        {localStorage.getItem('cached_projects') && projects === null
           && (
             <Box sx={{ textAlign: 'center' }}>
               <CircularProgress />
             </Box>
           )}
-        {/* If jwt token doesn't exist, no projects exist either  */}
-        {(!localStorage.getItem('jwt') || projects?.length === 0)
+        {/* If there is no project cached, no projects exist either. */}
+        {(!localStorage.getItem('cached_projects') || projects?.length === 0)
           && (
             <Alert severity="info">
-              You have not created any mappings yet.
+              You have not created any mappings yet. There are
               Create one by clicking the Plus Icon.
             </Alert>
           )}
@@ -197,7 +236,7 @@ function GeneMapperHome({ style, loggedIn }) {
                 ))}
             </div>
           )}
-        {deletedProjects.length > 0
+        {loggedIn && deletedProjects.length > 0
           && (
             <Box>
               <Typography variant="h6" sx={{ mt: 4, mb: 2 }}>Deleted Projects</Typography>
